fix: testrunner build
