+++ package.json
@@ -27,8 +27,12 @@
     "@babel/core": "^7.4.3",
     "@babel/preset-env": "^7.4.3",
     "@babel/preset-flow": "^7.0.0",
+    "@types/chai": "^4.1.7",
+    "@types/glob": "^7.1.1",
     "@types/mocha": "^5.2.6",
     "@types/node": "^11.13.2",
+    "@types/platform": "^1.3.2",
+    "@types/promise-polyfill": "^6.0.3",
     "appium-ios-simulator": "^3.10.0",
     "babel-eslint": "^10.0.1",
     "babel-loader": "^8.0.5",
@@ -45,7 +49,7 @@
     "express": "^4.16.4",
     "filenamify-url": "1.0.0",
     "flow-bin": "0.56.0",
-    "glob": "7.1.2",
+    "glob": "7.1.3",
     "html2canvas-proxy": "1.0.1",
     "jquery": "^3.4.0",
     "js-polyfills": "^0.1.42",
@@ -82,12 +86,13 @@
     "webpack-cli": "^3.3.0"
   },
   "scripts": {
-    "prebuild": "rimraf dist/",
-    "build2": "node scripts/create-reftest-list && npm run build:npm && npm run build:browser",
-    "build": "rollup -c rollup.config.ts",
+    "prebuild": "rimraf dist/ && rimraf build/ && mkdirp dist && mkdirp build",
+    "build": "rollup -c rollup.config.ts && npm run build:create-reftest-list && npm run build:testrunner",
+    "build:testrunner": "rollup -c tests/rollup.config.ts",
     "build:npm": "babel src/ -d dist/npm/ --plugins=dev-expression && replace-in-file __VERSION__ '\"$npm_package_version\"' dist/npm/index.js",
     "build:browser": "webpack",
     "build:reftest-result-list": "ts-node scripts/create-reftest-result-list.ts",
+    "build:create-reftest-list": "ts-node scripts/create-reftest-list.ts tests/reftests/ignore.txt build/reftests.ts",
     "build:reftest-preview": "webpack --config www/webpack.config.js",
     "release": "standard-version",
     "format": "prettier --single-quote --no-bracket-spacing --tab-width 4 --print-width 100 --write \"{src,www/src,tests,scripts}/**/*.js\"",
